refactor: logic of a game session is extracted from ConsoleHangman.java in SessionManager.java
